requirements.txt
fastapi==0.111.0
scikit-learn==1.3.2
uvicorn==0.40.0
httpx==0.26.0
pytest==9.0.2
mysql-connector-python
pandas